refactor: move mapping logic to converters.ts
